+++ src/pages/dashboard/settings/Plans.vue
@@ -7,12 +7,11 @@ import { Capacitor } from '@capacitor/core'
 import dayjs from 'dayjs'
 import { openCheckout } from '~/services/stripe'
 import { useMainStore } from '~/stores/main'
-import { findBestPlan, getCurrentPlanName, getPlanUsagePercent, getPlans, getTotalStats, useSupabase } from '~/services/supabase'
+import { findBestPlan, getCurrentPlanName, getPlanUsagePercent, getPlans, getTotalStats } from '~/services/supabase'
 import { useLogSnag } from '~/services/logsnag'
 import { openMessenger } from '~/services/chatwoot'
 import type { Database } from '~/types/supabase.types'
 import type { Stat } from '~/components/comp_def'
-import { bytesToGb } from '~/services/conversion'
 
 function openSupport() {
   openMessenger()
@@ -161,77 +160,6 @@ const hightLights = computed<Stat[]>(() => ([
     value: currentPlanSuggest.value?.name,
   },
 ]))
-
-const supabase = useSupabase()
-
-async function getUsage() {
-  const apps = await supabase.from('apps')
-    .select('app_id')
-    .eq('user_id', main!.user!.id!)
-
-  if (!apps.data)
-    return
-
-  const usage = await supabase.from('app_usage')
-    .select('mau, storage, bandwidth')
-    .gte('created_at', main.cycleInfo?.subscription_anchor_start)
-    .lte('created_at', main.cycleInfo?.subscription_anchor_end)
-    .in('app_id', apps.data.map(app => app.app_id))
-
-  const plan = plans.value.find(p => p.name === 'Pay as you go')!
-
-  let totalMau = 0
-  let totalStorage = 0
-  let totalBandwidth = 0
-
-  usage?.data?.forEach((item) => {
-    totalMau += item.mau
-    totalStorage += bytesToGb(item.storage)
-    totalBandwidth += bytesToGb(item.bandwidth)
-  })
-
-  const payg_base = {
-    mau: plan.mau,
-    storage: plan.storage,
-    bandwidth: plan.bandwidth,
-  }
-
-  const payg_units = {
-    mau: plan.mau_unit,
-    storage: plan.storage_unit,
-    bandwidth: plan.bandwidth_unit,
-  }
-
-  const basePrice = plan.price_m
-
-  const calculatePrice = (total: number, base: number, unit: number) => total <= base ? 0 : (total - base) * unit
-
-  const totalPrice = computed(() => {
-    const mauPrice = calculatePrice(totalMau, payg_base.mau, payg_units!.mau!)
-    const storagePrice = calculatePrice(totalStorage, payg_base.storage, payg_units!.storage!)
-    const bandwidthPrice = calculatePrice(totalBandwidth, payg_base.bandwidth, payg_units!.bandwidth!)
-    const sum = mauPrice + storagePrice + bandwidthPrice
-    return roundNumber(basePrice + sum)
-  })
-
-  return {
-    totalPrice,
-    totalMau,
-    totalBandwidth,
-    totalStorage,
-    plan,
-  }
-}
-
-function roundNumber(number: number) {
-  return Math.round(number * 100) / 100
-}
-
-const planUsage = ref<Awaited<ReturnType<typeof getUsage>>>()
-
-getUsage().then((res) => {
-  planUsage.value = res
-})
 </script>
 
 <template>
@@ -264,98 +192,6 @@ getUsage().then((res) => {
                 </p>
               </div>
             </div>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div v-if="currentPlan?.description.includes('payasyougo') && main.paying" class="px-4 pt-6 mx-auto max-w-7xl lg:px-8 sm:px-6">
-      <div class="sm:align-center sm:flex sm:flex-col">
-        <h1 class="text-5xl font-extrabold text-gray-900 sm:text-center dark:text-white">
-          {{ t('billing') }}
-        </h1>
-
-        <div class="my-2">
-          <div class="text-lg font-bold">
-            Monthly Active Users
-          </div>
-          <hr class="my-1 border-t-2 border-gray-300 opacity-70">
-          <div class="flex row justify-between mt-2">
-            <div>
-              Included in plan
-            </div>
-            <div class="font-semibold">
-              {{ planUsage?.plan.mau }}
-            </div>
-          </div>
-
-          <hr class="my-1 border-t border-gray-300 opacity-50">
-          <div class="flex row justify-between">
-            <div>
-              Used in period
-            </div>
-            <div class="font-semibold">
-              {{ planUsage?.totalMau }}
-            </div>
-          </div>
-        </div>
-
-        <div class="my-2">
-          <div class="text-lg font-bold">
-            Storage
-          </div>
-          <hr class="my-1 border-t-2 border-gray-300 opacity-70">
-          <div class="flex row justify-between mt-2">
-            <div>
-              Included in plan
-            </div>
-            <div class="font-semibold">
-              {{ planUsage?.plan.storage }} GB
-            </div>
-          </div>
-
-          <hr class="my-1 border-t border-gray-300 opacity-50">
-          <div class="flex row justify-between">
-            <div>
-              Used in period
-            </div>
-            <div class="font-semibold">
-              {{ planUsage?.totalStorage.toFixed(3) }} GB
-            </div>
-          </div>
-        </div>
-
-        <div class="my-2">
-          <div class="text-lg font-bold">
-            Bandwidth
-          </div>
-          <hr class="my-1 border-t-2 border-gray-300 opacity-70">
-          <div class="flex row justify-between mt-2">
-            <div>
-              Included in plan
-            </div>
-            <div class="font-semibold">
-              {{ planUsage?.plan.bandwidth }} GB
-            </div>
-          </div>
-
-          <hr class="my-1 border-t border-gray-300 opacity-50">
-          <div class="flex row justify-between">
-            <div>
-              Used in period
-            </div>
-            <div class="font-semibold">
-              {{ planUsage?.totalBandwidth.toFixed(3) }} GB
-            </div>
-          </div>
-        </div>
-        <hr class="my-1 border-t border-gray-300 opacity-30">
-        <div class="flex row justify-between">
-          <div>
-            Total estimated bill (Base + Overage)
-          </div>
-          <div class="font-semibold">
-            ${{ planUsage?.totalPrice }}
           </div>
         </div>
       </div>
